Complete EditTweet, Apply Login Security

+++ twitter-frontend/src/tweet/DeleteTweet.js
@@ -1,24 +1,32 @@
-import React from 'react'
 import { Button } from '@material-ui/core'
 import axios from 'axios'
+import React from 'react'
 import { useDispatch } from 'react-redux'
+import { tweetRedirect } from '../modules/redirect'
 import { getTweets } from './tweets'
-import { deleteTweetRedirect } from '../modules/redirect'
 
 function DeleteTweet(tweet) {
   const dispatch = useDispatch()
   const token = localStorage.token
-  const deleteTweet = async () => {
-    await axios.delete("http://localhost:8080/api/tweets/" + tweet.id, {
-      headers:
-      {
-        Authorization: 'Bearer ' + token
-      }
-    }).then(
-      dispatch(getTweets()),
-      dispatch(deleteTweetRedirect())
-    )
+
+  const deleteTweet = async (e) => {
+    if (e.keyCode === 13 || e.button === 0) {
+      axios.delete("http://localhost:8080/api/tweets/" + tweet.id, {
+        headers:
+        {
+          Authorization: 'Bearer ' + token
+        }
+      }).then(
+        setTimeout(
+          function() {
+            dispatch(getTweets())
+          },500
+        ),
+        dispatch(tweetRedirect())
+      )
+    }
   }
+
 
   return (
     <div>
